training/requirements.txt
# RetailSync AI - Training Environment

## 📦 Required Dependencies

tensorflow>=2.15.0
tensorflowjs>=4.14.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
